+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## [0.15.1] - 2024-03-05
+
+- Fix codegen for custom grad functions calling their respective forward functions
+- Fix custom grad function handling for functions that have no outputs
+- Fix issues when `wp.config.quiet = True`
 
 ## [0.15.0] - 2024-03-04